Moved CurrencyRepository into a Currency directory
