+++ IdelPog/Assets/Scripts/IdelPog/Repository/Currency.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 163ede81bf6a472da7374a4d7b402795
+timeCreated: 1735937561
