+++ IdelPog/Assets/Scripts/Tests/Service/TestableCurrencyService.cs
@@ -1,4 +1,5 @@
 ﻿using IdelPog.Repository;
+using IdelPog.Repository.Currency;
 using IdelPog.Service.Currency;
 
 namespace Tests.Service
